Rating connected to the api

+++ src/backend/src/Wemcy.RecipeApp.Backend/Model/Ingredient.cs
@@ -6,6 +6,6 @@ namespace Wemcy.RecipeApp.Backend.Model;
 public record Ingredient
 {
     public required string Name { get; set; }
-    public required double Quantity { get; set; }
+    public required float Quantity { get; set; }
     public required string UnitOfMeasurement { get; set; }
 }
